training/requirements.txt
pandas>=1.3.5
pycaret[full]>=2.3.10
boto3>=1.26.0
python-dotenv>=0.21.0
scikit-learn>=1.0.2
joblib>=1.1.0
pyarrow>=7.0.0